Soft-deprecate FileAttr::padding
This was added in 7f677ce34642870b36b2bb73c952762b42c26eac, but I think it
was a mistake.  AFAICS the padding is required by the FUSE ABI, but isn't
actually used, so it shouldn appear in `fuse_abi::fuse_attr`, but not in
in our `FileAttr` API.

Removing it now would be a breaking change, so I'll just ignore it and
document that it's unused.  It can be removed for 0.8.