Remove unsafe from the request parsing code
Instead we use `zerocopy::FromBytes`.

Previously the code did not ensure alignment, which could lead to undefined
behaviour.  Now we do check alignment I've had to fix it in a few places
using `AlignedData<[u8; sz]>` as a buffer in tests.  Unfortunately we can't
use `Box<AlignedData<u8; BUFFER_SIZE>>` as our main buffer because it's too
large to create on the stack and Rust currently offers no way to allocate
it directly on the heap.  Instead we have to fiddle with offsets for the
`Vec`.