Channel safety: Have seperate `Mount` and `Channel` objects
This splits `Channel` in two.  `Channel` now only encapsulates (and owns)
the FUSE device.  It is no longer responsible for unmounting the
filesystem.  Unmounting falls to a new, backend-specific, struct `Mount`.

The `Mount` belongs to session.  This way the `Session` can drop the mount
without also closing the fuse device file. This is important as we don't
want to close an FD that is still in use as we can cause the equivalent
of use-after-free bugs†.  When a filesystem is unmounted the devices
return `ENODEV`, so should naturally exit as well in an orderly fashion
and without races.

Ownership is now a lot more streightforward.  In particular the libfuse3
`fuse_session` lifecycle was a little confused before, while now we use
RAII in a streightforward fashion to ensure things are cleaned up
correctly, and not cleaned up twice.  The previous implementation had a
bug where it would never be freed because `fuse_session_unmount` would
only be called on null pointers.

The fuse device objects are plain `std::fs::File`s with ownership over the
underlying FD.  Note: although they are Files, we still implement our own
reading and writing via `libc::read` and `libc::writev` due to concerns
about the guarantees that `std::fs::File`'s `AsyncRead` and `AsyncWrite`
provides.  We can still revisit this decison in the future.

Generally I've followed the principle that it's a good idea to encapsulate
an FD as soon as we can for greater confidence that it will get cleaned
up correctly.

† Note: there is still the possbility of writing to a file descriptor
  after it has been closed.  I'll fix this in the future.