ll::Request: `FilenameInDir` consistency
I've documented invariants here, and pulled this up a level in the mod
hierarchy.

I've used `src()` and `dest()` everywhere for consistency rather than the
mismash of terms that were here before.  `dest` means that an entry should
be created at this path.  `path()` is still used in some cases.