Enum API: Add `mount3` and `serve_sync` APIs
Can be used to mount and serve requests for filesystems implemented in
terms of the enum API.

Note: This is a little different to the `mount` and `mount2` APIs in that
we've split out mounting from actually serving.  This is neat as it removes
the need for a separate `spawn_mount` API.  If the user wants to run their
filesystem from a different thread they can just call `serve_sync` in that
thread.

It also provides the flexibility to do other work between `mount` and
`serve` - such as readyness notification.