Make the use of fusermount(1) explicit
There are multiple ways to mount FUSE filesystems. Processes with
elevated privileges (CAP_SYS_ADMIN) can use the mount(2) syscall
directly or via a suitable helper. libfuse also installs a helper
program, fusermount(1), which is setuid root, to perform this mounting
operation.

Previously, we would attempt the syscall, and then fallback to trying
fusermount (if we can find it) automatically. This is a little bit too
magical, so it's better to make the option explicit.