Don't doubly unmount the file system when using libfuse
fuse_session_destroy already includes an unmount, so it's redundant do
call both that and fuse_session_unmount.  It's dangerous even, because
it could race with a different process using the same mountpoint, and
unmount that process's file system.

This change will be more important after the next libfuse release, which
will fix a similar double-unmount problem: 1287[^1].

Both this commit and [^1] are necessary to eliminate
"unmounting /tmp/mnt failed: Invalid argument" warnings that appeared
beginning with libfuse 3.17.0.

[^1]: https://github.com/libfuse/libfuse/pull/1287